+++ src/pipe/connector.h
@@ -190,8 +190,8 @@ dt_connector_channels(const dt_connector_t *c)
 static inline size_t
 dt_connector_bufsize(const dt_connector_t *c, uint32_t wd, uint32_t ht)
 {
-  if(c->format == dt_token("bc1")) return wd/4*ht/4 * 8;
-  if(c->format == dt_token("yuv")) return wd*ht * 2; // XXX ??? * 2/3 is not enough here.
+  if(c->format == dt_token("bc1")) return wd/4*(uint64_t)ht/4 * (uint64_t)8;
+  if(c->format == dt_token("yuv")) return wd*(uint64_t)ht * (uint64_t)2;
   if(c->format == dt_token("geo")) return 14*sizeof(int16_t)*ht/3; // 14 values per triangle
   const int numc = dt_connector_channels(c);
   const size_t bpp = dt_connector_bytes_per_channel(c);
